Mouse input

See merge request profun/manylands!4

+++ src/Curve_stats.h
@@ -31,9 +31,9 @@ struct Curve_stats
         Min_and_max w;
     };
 
-    double min_speed, max_speed;
+    float min_speed, max_speed;
     // These two vectors should be of the size of the curve (number of points)
-    std::vector<double> speed;
+    std::vector<float> speed;
     std::vector<std::string> dimensionality;
     // Vectors bellow have the size depending from number of switches
     std::vector<size_t> switches_inds;
